+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for utf8 encoding to `*_to_atom` and `atom_to_*` functions
 - `binary_to_atom/1` and `atom_to_binary/1` that default to utf8 (they were introduced with OTP23)
 - Added Pico cmake option `AVM_WAIT_BOOTSEL_ON_EXIT` (default `ON`) to allow tools to use automated `BOOTSEL` mode after main application exits
+- Use UTF-8 encoding for atoms when using `erlang:term_to_binary/1`, in conformance with OTP-26
 
 ### Fixed
 
@@ -46,7 +47,6 @@ used)
 - New atom table, which uses less memory, has improved performances and better code.
 - SPI: when gpio number is not provided for `miso` or `mosi` default to disabled
 - Change port call tuple format to the same format as gen_server, so casts can be supported too
-- Use UTF-8 encoding for atoms when using `erlang:term_to_binary/1,2`, in conformance with OTP-26
 
 ### Fixed